Renaming test bodies
