+++ grails-app/services/uk/ac/ox/softeng/maurodatamapper/plugins/digitalobjectidentifiers/profile/DigitalObjectIdentifiersProfileProviderService.groovy
@@ -30,7 +30,7 @@ class DigitalObjectIdentifiersProfileProviderService extends JsonProfileProvider
 
     @Override
     String getDisplayName() {
-        'DOI DataCite Dataset Schema'
+        'Digital Object Identifiers DataCite Dataset Schema'
     }
 
     @Override
